+++ apps/application/flow/step_node/__init__.py
@@ -24,7 +24,7 @@ from .search_dataset_node import *
 from .speech_to_text_step_node import BaseSpeechToTextNode
 from .start_node import *
 from .text_to_speech_step_node.impl.base_text_to_speech_node import BaseTextToSpeechNode
-from .variable_assign import BaseVariableAssignNode
+from .variable_assign_node import BaseVariableAssignNode
 
 node_list = [BaseStartStepNode, BaseChatNode, BaseSearchDatasetNode, BaseQuestionNode,
              BaseConditionNode, BaseReplyNode,
